requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
pandas